Revert "Remove session and security roles from freemarker vars"
