+++ orcid-core/src/main/java/org/orcid/core/manager/v3/SourceManager.java
@@ -2,6 +2,7 @@ package org.orcid.core.manager.v3;
 
 import org.orcid.jaxb.model.v3.rc2.common.Source;
 import org.orcid.persistence.jpa.entities.ProfileEntity;
+import org.orcid.persistence.jpa.entities.SourceAwareEntity;
 import org.orcid.persistence.jpa.entities.SourceEntity;
 
 /**
