+++ orcid-web/src/main/resources/freemarker/admin/admin_actions.ftl
@@ -2,8 +2,7 @@
 <div class="row">
     <!-- Left menu bar -->  
     <div class="col-md-3 col-sm-12 col-xs-12 lhs padding-fix">
-        <#include "/includes/ng2_templates/id-banner-ng2-template.ftl"/>
-        <id-banner-ng2> </id-banner-ng2>
+        <#include "../includes/id_banner.ftl"/>
     </div>
     <!-- Right menu bar -->
     <div class="col-md-9 col-sm-12 col-xs-12 admin-options">
