+++ orcid-web/src/main/resources/freemarker/admin/manage_members.ftl
@@ -113,8 +113,7 @@
 <div class="row">
 	<!-- Left menu bar -->	
 	<div class="col-md-3 col-sm-12 col-xs-12 lhs padding-fix">
-	    <#include "/includes/ng2_templates/id-banner-ng2-template.ftl"/>
-	    <id-banner-ng2> </id-banner-ng2>
+		<#include "../includes/id_banner.ftl"/>
 	</div>
