+++ orcid-web/src/main/resources/freemarker/admin_menu.ftl
@@ -1,5 +1,4 @@
 <div class="lhs">
-    <#include "/includes/ng2_templates/id-banner-ng2-template.ftl"/>
-    <id-banner-ng2> </id-banner-ng2>
+	<#include "includes/id_banner.ftl"/>
 </div>
 <div class="lhs override"></div>
