+++ orcid-web/src/main/resources/freemarker/developer_tools/developer_tools.ftl
@@ -1,8 +1,7 @@
 <@public nav="developer-tools">
 	<div class="row">
 		<div class="col-md-3 lhs col-sm-12 col-xs-12 padding-fix">
-		    <#include "/includes/ng2_templates/id-banner-ng2-template.ftl"/>
-            <id-banner-ng2> </id-banner-ng2>
+			<#include "../includes/id_banner.ftl"/>
 		</div>
 		<div class="col-md-9 col-sm-12 col-xs-12 developer-tools">
 		    <h1 id="manage-developer-tools">
