+++ orcid-web/src/main/resources/freemarker/manage.ftl
@@ -22,8 +22,7 @@
 <div class="row">
     <div class="col-md-3 col-sm-12 col-xs-12 padding-fix lhs">
         <div class="lhs">
-            <#include "/includes/ng2_templates/id-banner-ng2-template.ftl"/>
-            <id-banner-ng2> </id-banner-ng2>
+            <#include "includes/id_banner.ftl"/>
         </div>  
     </div>
     <!-- Right side -->
