+++ orcid-web/src/main/resources/freemarker/member_developer_tools.ftl
@@ -2,8 +2,7 @@
 
 <div class="row developer-tools">
 	<div class="col-md-3 col-sm-12 col-xs-12">
-	    <#include "/includes/ng2_templates/id-banner-ng2-template.ftl"/>
-        <id-banner-ng2> </id-banner-ng2>
+		<#include "includes/id_banner.ftl"/>
 	</div>
 	<div class="col-md-9 col-sm-12 col-xs-12 margin-top-box-mobile">
 		<#include "/includes/ng2_templates/client-edit-ng2-template.ftl">
