+++ orcid-web/src/main/resources/freemarker/self_service.ftl
@@ -8,8 +8,7 @@
     
     <div class="row">
         <div class="col-md-3 lhs col-sm-12 col-xs-12 padding-fix">
-            <#include "/includes/ng2_templates/id-banner-ng2-template.ftl"/>
-            <id-banner-ng2> </id-banner-ng2>
+            <#include "includes/id_banner.ftl"/>
         </div>
         <self-service-ng2></self-service-ng2>
     </div>
