+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/modules/adminActions/adminActions.component.ts
@@ -20,7 +20,8 @@ import { CommonService }
     
 @Component({
     selector: 'admin-actions-ng2',
-    template:  scriptTmpl("admin-actions-ng2-template")
+    template:  scriptTmpl("admin-actions-ng2-template"),
+    providers: [CommonService]
 })
 export class AdminActionsComponent implements AfterViewInit, OnDestroy, OnInit {
     private ngUnsubscribe: Subject<void> = new Subject<void>();
