+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/modules/affiliation/affiliation.component.ts
@@ -34,7 +34,8 @@ import { CommonService }
     
 @Component({
     selector: 'affiliation-ng2',
-    template:  scriptTmpl("affiliation-ng2-template")
+    template:  scriptTmpl("affiliation-ng2-template"),
+    providers: [CommonService]
 })
 export class AffiliationComponent implements AfterViewInit, OnDestroy, OnInit {
     @Input() publicView: any;
