+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/modules/common/common.ts
@@ -288,6 +288,7 @@ import { PersonService }
             BlogService,
             ClaimService,
             ClientService,
+            CommonService,
             ConsortiaService,
             DiscoService,
             EmailFrequencyService,
