+++ orcid-web/src/main/webapp/static/javascript/ng1Orcid/app/modules/header/header.component.ts
@@ -14,18 +14,12 @@ import { FormsModule }
 
 import { Observable, Subject, Subscription } 
     from 'rxjs';
-    
-import { takeUntil, shareReplay } 
+import { takeUntil } 
     from 'rxjs/operators';
-    
+
 import { NotificationsService } 
     from '../../shared/notifications.service.ts'; 
 
-import { CommonService } 
-    from '../../shared/common.service.ts';
-    
-import { FeaturesService }
-    from '../../shared/features.service.ts';
 
 @Component({
     selector: 'header-ng2',
@@ -44,16 +38,9 @@ export class HeaderComponent implements AfterViewInit, OnDestroy, OnInit {
     secondaryMenuVisible: any;
     settingsVisible: boolean;
     tertiaryMenuVisible: any;
-    userInfo: any;
-    isOauth: boolean = false;
-    isPublicPage: boolean = false;
-    profileOrcid: string = null;
-    showSurvey = this.featuresService.isFeatureEnabled('SURVEY');
-    
+
     constructor(
-        private notificationsSrvc: NotificationsService,
-        private featuresService: FeaturesService,
-        private commonSrvc: CommonService
+        private notificationsSrvc: NotificationsService
     ) {
         this.conditionsActive = false;
         this.filterActive = false;
@@ -65,36 +52,8 @@ export class HeaderComponent implements AfterViewInit, OnDestroy, OnInit {
         this.secondaryMenuVisible = {};
         this.settingsVisible = false;
         this.tertiaryMenuVisible = {};
-        const urlParams = new URLSearchParams(window.location.search);
-        this.isOauth = (urlParams.has('client_id') && urlParams.has('redirect_uri'));
-        var orcidRegex = /^(\d{4}-){3}\d{3}[\dX]$/;
-        var path = window.location.pathname;
-        path = path.substring(path.lastIndexOf('/') + 1);
-        this.isPublicPage = orcidRegex.test(path);
-        if(this.isPublicPage) {            
-            this.profileOrcid = path;
-            this.userInfo = this.commonSrvc.getPublicUserInfo(this.profileOrcid)
-            .subscribe(
-                data => {
-                    this.userInfo = data;
-                },
-                error => {
-                    console.log('header.component.ts: unable to fetch publicUserInfo', error);
-                } 
-            );
-        } else {
-            this.userInfo = this.commonSrvc.userInfo$
-            .subscribe(
-                data => {
-                    this.userInfo = data;                
-                },
-                error => {
-                    console.log('header.component.ts: unable to fetch userInfo', error);
-                    this.userInfo = {};
-                } 
-            );
-        }      
     }
+
     
     filterChange(): void {
         this.searchFilterChanged = true;
@@ -214,6 +173,6 @@ export class HeaderComponent implements AfterViewInit, OnDestroy, OnInit {
 
     ngOnInit() {
         this.onResize(); 
-        this.headerSearch.searchOption = 'registry';         
+        this.headerSearch.searchOption = 'registry'; 
     }; 
 }
